+++ packages/adp-tooling/src/types.ts
@@ -199,12 +199,20 @@ export interface CodeExtChange extends CommonChangeProperties {
     };
 }
 
-export const CUSTOMER_BASE = 'CUSTOMER_BASE';
-
 export const enum TemplateFileName {
     Fragment = 'fragment.xml',
     Controller = 'controller.ejs',
     Annotation = 'annotation.xml'
+}
+
+export const enum FlexLayer {
+    CUSTOMER_BASE = 'CUSTOMER_BASE',
+    VENDOR = 'VENDOR'
+}
+
+export const enum NamespacePrefix {
+    CUSTOMER = 'customer.',
+    EMPTY = ''
 }
 
 export const enum HttpStatusCodes {
@@ -322,30 +330,28 @@ export interface ComponentUsagesData {
 }
 
 export interface NewModelData {
-    projectData: AdpProjectData;
-    timestamp: number;
-    annotation: {
-        /** Name of the OData annotation data source. */
-        dataSourceName: string;
-        /** Optional URI of the OData annotation data source. */
-        dataSourceURI?: string;
-        /** Optional settings for the OData annotation. */
-        settings?: string;
-    };
-    service: {
-        /** Name of the OData service. */
-        name: string;
-        /** URI of the OData service. */
-        uri: string;
-        /** Name of the OData service model. */
-        modelName: string;
-        /** Version of OData used. */
-        version: string;
-        /** Settings for the OData service model. */
-        modelSettings: string;
-    };
-    /** Indicates whether annotation mode is added. */
+    variant: DescriptorVariant;
+    answers: NewModelAnswers;
+}
+
+export interface NewModelAnswers {
     addAnnotationMode: boolean;
+    /** Name of the OData service. */
+    name: string;
+    /** URI of the OData service. */
+    uri: string;
+    /** Name of the OData service model. */
+    modelName: string;
+    /** Version of OData used. */
+    version: string;
+    /** Settings for the OData service model. */
+    modelSettings: string;
+    /** Name of the OData annotation data source. */
+    dataSourceName: string;
+    /** Optional URI of the OData annotation data source. */
+    dataSourceURI?: string;
+    /** Optional settings for the OData annotation. */
+    annotationSettings?: string;
 }
 
 export interface DataSourceData {