+++ packages/fiori-elements-writer/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @sap-ux/fiori-elements-writer
 
+## 1.0.8
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.25.4
+-   @sap-ux/fiori-generator-shared@0.3.4
+-   @sap-ux/odata-service-writer@0.22.1
+-   @sap-ux/ui5-application-writer@1.1.2
+-   @sap-ux/ui5-test-writer@0.4.0
+
 ## 1.0.7
 
 ### Patch Changes